Extracted infos parser from ffmpeg reader to own class
